create many-to-many relation between VideoGame and Platform models

+++ db/migrate/20240227132341_create_video_games.rb
@@ -4,7 +4,7 @@ class CreateVideoGames < ActiveRecord::Migration[7.0]
       t.string :name
       t.date :release_date
       t.string :studio
-      t.integer :ratings
+      t.float :ratings
 
       t.timestamps
     end
